+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `fill_data` function
 - `random_hex_color_gen` function
 - `color`,`bgcolor` and `projection` parameters random mode
 ### Changed